+++ src/models/model_answer.py
@@ -108,7 +108,7 @@ class ModelAnswer:
     # Marking content
     answer_text: str = ""
     guideline_text: Optional[str] = None
-    max_marks: Optional[int] = None
+    max_marks: Optional[float] = None
 
     # Paper metadata
     module_code: Optional[str] = None
